improve labels in image info table + add game count to image info in replays
